use swing instead of javafx

+++ komap-core/src/main/kotlin/nice/fontaine/models/Coordinates.kt
@@ -8,4 +8,4 @@ data class TCoord(var x: Int, var y: Int)
 /**
  * Canvas Coordinate
  */
-data class CCoord(var x: Double, var y: Double)
+data class CCoord(var x: Int, var y: Int)
